+++ Assets/Programming/HGS/Scripts/Resource/ResourceData.cs.meta
